Working on last commit debug (part 2)
- Pick function added
- Effects now affect Ability
- Bonus effects added to Ability now decrease duration
- InteractiveType was fixed
- Search with Manager fixed

+++ Source/Workshop/ActorClasses/InteractiveObjects/InteractiveCharacter.cpp
@@ -5,7 +5,7 @@
 
 AInteractiveCharacter::AInteractiveCharacter()
 {
-  InteractiveType = EInteractiveType::Character;
+  InteractiveType = static_cast<int32>(EInteractiveType::Character);
 
   CharacterPresentation = CreateDefaultSubobject<UPaperFlipbookComponent>(TEXT("CharacterSprite"));
   CharacterPresentation->SetupAttachment(RootComponent);
